v1/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
dnspython==2.7.0
dotenv-python==0.0.1
eval-type-backport==0.2.2
feedparser==6.0.11
gnews==0.4.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jsonpath-python==1.0.6
loguru==0.7.3
mistralai==1.5.0
mypy-extensions==1.0.0
pydantic==2.10.6
pydantic-core==2.27.2
python-dateutil==2.9.0.post0
randfacts==0.22.1
requests==2.32.3
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
typing-extensions==4.12.2
typing-inspect==0.9.0
urllib3==2.3.0
